test: Extract Radio.Metrics.Tests from scattered test projects
Move 17 metrics tests into a dedicated Radio.Metrics.Tests project:
- 3 model/enum tests from Radio.Core.Tests/Metrics/
- 2 data layer tests from Radio.Infrastructure.Tests/Metrics/
- Remove unused Radio.Metrics reference from Radio.Core.Tests
